requirements.txt
tensorflow-gpu
matplotlib == 3.0.2
numpy
keras
xarray
pandas
basemap
